perf: implement redirect path (see description)
at first there was some issues in the login page where sometimes it just stays in login page  after user logged in. it was some issues with the navigate(-1). so i added a redirectpath state in usercontext to get where the user was before. so we can use navigate(redirectPath) wihich fixed the problem